Implement some of the state hooks for the homepage
